legacy/backend/pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fastapi-backend"
version = "0.1.0"
description = "FastAPI 백엔드 서비스"
requires-python = ">=3.8"
dependencies = [
    "debugpy>=1.6.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.23.2",
    "pydantic>=2.4.2",
    "psycopg2-binary>=2.9.9",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.1.0",
    "python-dotenv>=1.0.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-api-python-client>=2.100.0"
]
